+++ packages/providers/google/src/embedding.ts
