+++ packages/providers/google/src/index.ts
@@ -1,6 +1,6 @@
-export * from "./base";
-
+export * from "./embedding";
 export * from "./live";
+export * from "./llm";
 export {
   GoogleStudio,
   Modality,
@@ -8,5 +8,3 @@ export {
 } from "./studio/index.js";
 export * from "./types";
 export * from "./utils";
-
-export * from "./GeminiEmbedding";
